add the converted states to the BloodDudeBase class.

+++ source/games/blood/src/blood.cpp
@@ -803,8 +803,8 @@ DEFINE_PROPERTY(aistate, SSIIGGGGs, CoreActor)
 
 	char* endp = (char*)"";
 	if (*seq == 0) state.sprite = 0;
-	else if (*seq == '#') state.sprite = strtoull(seq + 1, &endp, 10) | 0x10000000;
-	else if (*seq == '+') state.sprite = strtoull(seq + 1, &endp, 10) | 0x20000000;
+	else if (*seq == '#') state.sprite = (int)strtoull(seq + 1, &endp, 10) | 0x10000000;
+	else if (*seq == '+') state.sprite = (int)strtoull(seq + 1, &endp, 10) | 0x20000000;
 
 	state.Label = label;
 	state.Type = type;
